+++ R/zzz.R
@@ -2,14 +2,16 @@
 
 
 #' @importFrom grDevices colorRampPalette
-#' @importFrom methods is
-#' @importFrom stats cmdscale dist sd
+#' @importFrom stats cmdscale dist sd aggregate
 #' @importFrom utils packageVersion
 #' @importFrom utils write.csv
 #' @importFrom graphics par lines axis
 #' @importFrom grDevices pdf
 #' @importFrom grDevices dev.off
+#' @importFrom methods as is
 #' @importFrom utils head
+
+
 
 utils::globalVariables(c("."))
 
@@ -33,7 +35,7 @@ if (getRversion() >= "2.15.1") {
       "orig.ident", "parallel_paramSweep", "productive", "proportion",
       "receptor_subtype", "receptor_type", "row_id", "sample", "scores",
       "stat_method", "total", "value", "values", "variable", "x", "top", "name", "object_temp","method",
-      "overall_rank", "asw", "ch", "db", "ari", "nmi", "overall","Feature", "Proportion"
+      "overall_rank", "asw", "ch", "db", "ari", "nmi", "overall","Feature", "Proportion","variance"
     )
   )
 }